Add dialog and enhance functionality in main.go
This commit introduces a dialog for a user input and redesigns the functionality of the current system. It starts by adding an import for dialog from fyne library. Three buttons were added for advanced interactions: one for admitting that the user is hurt, one for the return of the user, and a button to restart the process. Additionally, the function simulator was removed and replaced by two functions, beBlack and beWhite, to create an output message about the status of the user relationship.